The all operations that load from 8 bit registers to/from 16 bit registers are removed.
That's because I misinterpret the original spec. of the Gameboy. To be
more specific, I thought operations like:

1- LD A, (HL)
2 - LD (nn), A

(1) load 16 register value (HL) to A and (2) value of register A to nn

but the correct behaviour was:

(1) load the value at memory in the index that value of HL and put it
into register A, (2) read the value of register A, and load it into
memory at the index nn ie. mem[nn] = A.

So none of the 8 bit registers put directly data on the 16 bit registers and
vice versa. So all of the operators= overloads were a mistake as a result of
misinterpreation mentioned above.