Add memory as a member to gain write/read access to cpu.
The other options was use a bus to write to/read from memory which is
implemented in bus branch. The bus does nothing other than
forward operation to memory, but it's considered a better practice in
long run. I have really no idea which approch is more
useful. So implemented them both way.