Change std::uint8_t and std::uint16 types with their 'fast' counterparts
I wonder what are the fast integer types a useful for and make search
about them. Then saw is StackOverflow* question, and had conclusion that
these are exactly what I'm looking for because u8 and u16 types
definitions represent internal data of the registers (reg8 and reg16) of
the cpu. Because registers suppose to fast:smile:, I changed the
corresponding changes.

https://stackoverflow.com/questions/46446453/what-is-a-fast-integer-what-defines-how-fast-an-integer-is