Initial implementation of the clock.
The clock class for cpu is added. The shared the draft of the code, one
emudev Discord Server, lots of nice and humble guys in there said this
simply will be non-efficient, because of the tick member request thread
to sleep. And this member called and 100'000 times (if it's not more)
and such a big number of syscal will be inefficient.

The current practice is using T-cycles instead M-cycles and doing that
to at the end of each frame. This method known to be successful. But
I decided to follow my own idea, I don't even know how this will be
turn out. So yeah, I'll try and see.