Instruction loop: Opcodes 0x30-0x03f put in their place.
This table is used: https://izik1.github.io/gbops/

DEC(const reg16 rr) instruction changed to DEC(). It's now doesn't take
argument because it's only used HL register. This is a result of another
misinterpretation of mine on Gameboy specs. I forget that HL is
a spesific register and interpret is generic represetation of 16 bit
register, like H is for High and L is for Low.

dummy parameters deleted from one overload of INC and DEC instructions. The primarily
reason I add that, to avoid redefinition.