Implemented: adc instructions.
Add with carry instruction overloads are implemented.
This instruction add value to A register with carry flags.
if the carry flag set, it adds plus one to A. What it does actually
summurises as:

A = A + r8 + c_flag     (1 machine cycle)
A = A + [HL] + c_flag   (2 m. cycle)
A = A + n8 + c_flag     (2 m. cycle)

- r8 a 8 bit cpu register (just like A itself)

- [byte] is a byte in memory indexed by HL (H for high, and L for low,
which can be any 16 bit register)

- n8 is a basically 8 bit unsigned value but put inside a class to avoid
  accidental integer promotions.